Backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib
email-validator
redis
python-multipart
aiofiles
filetype
bcrypt